Project Management snapshot
This is a PDF snapshot of an Excel file of a Military Operation from my time in the Marine Corps Air Wing. I have chosen to share an "Early Snapshot" of this military operation. The finished project information is sensitive information requiring a security clearance. I can assure you the project succeeded with 100% mission effectiveness and roughly a 12.5% cost savings in the operational budget. 
I have many more instances of project management experience using different software, however much of that information is the property of the respective employer, making it difficult to display.