Dataset Using JASP-Stats (R,C++,QML,jQuery,java)
in this folder, there are 5 files. 
2 CSV Files that are the raw datasets.
2 Jasp Files which can be opened if you have the software.
1 Word document full of Screenshots, Explanations, Questions & Answers.

I chose to upload this as a sample of some of my statistical analysis and visualization experience I hold. I have more experience than what is shown in this file. This is a small sample. The discussed topic is Correlation and Regression. even though the dataset is biological, this can easily be transferred into many different settings including Business decisions.